requirements.txt
requests~=2.27.1
numpy~=1.22.3
scikit-learn~=1.0.2
Django~=4.0.4
beautifulsoup4~=4.11.1
Pillow~=9.1.0
google-cloud-secret-manager~=2.10.0
mysqlclient==2.1.0